Refactor, batched subscriptions, boosts, many fixes

+++ bija/db.py
@@ -11,6 +11,7 @@ from sqlalchemy.sql.functions import coalesce
 from sqlalchemy.dialects.sqlite import insert as sqlite_upsert
 
 from bija.args import args
+from bija.helpers import is_hex_key
 from bija.models import *
 
 DB_ENGINE = create_engine("sqlite:///{}.sqlite".format(args.db), echo=False, poolclass=SingletonThreadPool, pool_size=10)
@@ -81,46 +82,34 @@ class BijaDB:
         new_unfollows = []
         is_mine = False
         for pk, item in lists.items():
-            following = self.get_following_pubkeys(pk)
-            new = set(item['pks']) - set(following)
-            removed = set(following) - set(item['pks'])
-            for pk2 in new:
-                inserts.append({'pk_1':pk, 'pk_2':pk2})
-            self.session.query(Follower).filter(Follower.pk_1 == pk).filter(Follower.pk_2.in_(removed)).delete()
-            self.session.commit()
+            if is_hex_key(pk):
+                following = self.get_following_pubkeys(pk)
+                new = set(item['pks']) - set(following)
+                removed = set(following) - set(item['pks'])
+                for pk2 in new:
+                    if is_hex_key(pk2):
+                        inserts.append({'pk_1':pk, 'pk_2':pk2})
+                self.session.query(Follower).filter(Follower.pk_1 == pk).filter(Follower.pk_2.in_(removed)).delete()
+                self.session.commit()
 
-            if pk == my_pk:
-                is_mine = True
-            if my_pk in new:
-                new_follows.append(pk)
-            if my_pk in removed:
-                new_unfollows.append(pk)
-
-        stmt = sqlite_upsert(Follower).values(inserts)
-        stmt = stmt.on_conflict_do_nothing(
-            index_elements=[Follower.pk_1, Follower.pk_2]
-        )
-        try:
-            self.session.execute(stmt)
-            self.session.commit()
-        except SQLAlchemyError as e:
-            print(str(e.__dict__['orig']))
+                if pk == my_pk:
+                    is_mine = True
+                if my_pk in new:
+                    new_follows.append(pk)
+                if my_pk in removed:
+                    new_unfollows.append(pk)
+        if len(inserts) > 0:
+            stmt = sqlite_upsert(Follower).values(inserts)
+            stmt = stmt.on_conflict_do_nothing(
+                index_elements=[Follower.pk_1, Follower.pk_2]
+            )
+            try:
+                self.session.execute(stmt)
+                self.session.commit()
+            except SQLAlchemyError as e:
+                print(str(e.__dict__['orig']))
 
         return new_follows, new_unfollows, is_mine
-
-    # def add_contact_list(self, public_key, keys: list):
-    #     following = self.get_following_pubkeys(public_key)
-    #     new = set(keys) - set(following)
-    #     removed = set(following) - set(keys)
-    #     a = []
-    #     for pk in new:
-    #         a.append(' ("{}","{}")'.format(public_key, pk))
-    #     if len(new) > 0:
-    #         sql = 'INSERT INTO follower (pk_1, pk_2) VALUES {}'.format(','.join(a))
-    #         DB_ENGINE.execute(text(sql))
-    #     self.session.query(Follower).filter(Follower.pk_1==public_key).filter(Follower.pk_2.in_(removed)).delete()
-    #     self.session.commit()
-    #     return list(new), list(removed)
 
     def get_last_contacts_upd(self, public_key):
         result = self.session.query(Event.ts) \
@@ -140,8 +129,11 @@ class BijaDB:
             self.session.query(Follower).filter(Follower.pk_1==my_pk).filter(Follower.pk_2==their_pk).delete()
         self.session.commit()
 
-    def get_following_pubkeys(self, public_key):
-        keys = self.session.query(Follower).filter_by(pk_1=public_key).all()
+    def get_following_pubkeys(self, public_key, start=None, end=None):
+        if start is not None and end is not None:
+            keys = self.session.query(Follower).filter_by(pk_1=public_key)[start:end]
+        else:
+            keys = self.session.query(Follower).filter_by(pk_1=public_key).all()
         out = []
         for k in keys:
             out.append(k.pk_2)
@@ -571,18 +563,8 @@ class BijaDB:
             return out
         return None
 
-    # def get_notes_by_id_list(self, note_ids):
-    #     return self.session.query(
-    #         Note.id,
-    #         Note.public_key,
-    #         Note.content,
-    #         Note.created_at,
-    #         Note.members,
-    #         Note.media,
-    #         Profile.name,
-    #         Profile.display_name,
-    #         Profile.pic,
-    #         Profile.nip05).join(Note.profile).filter(Note.id.in_(note_ids)).all()
+    def get_notes_by_id_list(self, note_ids):
+        return self.session.query(Note).filter(Note.id.in_(note_ids)).all()
 
     def get_unseen_message_count(self):
         return self.session.query(PrivateMessage) \
@@ -682,6 +664,10 @@ class BijaDB:
 
     def set_note_seen(self, note_id):
         self.session.query(Note).filter(Note.id == note_id).update({'seen': True})
+        self.session.commit()
+
+    def set_note_boosted(self, note_id):
+        self.session.query(Note).filter(Note.id == note_id).update({'shared': True})
         self.session.commit()
 
     def search_profile_name(self, name_str):
@@ -861,7 +847,7 @@ class BijaDB:
         self.session.commit()
 
     def get_alerts(self):
-        return self.session.query(Alert).order_by(Alert.seen.asc()).order_by(Alert.ts.desc()).limit(20).all()
+        return self.session.query(Alert).order_by(Alert.seen.asc()).order_by(Alert.ts.desc()).limit(50).all()
 
     def get_unread_alert_count(self):
         return self.session.query(Alert).filter(Alert.seen == 0).count()
